+++ Frends.SFTP.DownloadFiles/Frends.SFTP.DownloadFiles/Definitions/FileEncoding.cs
@@ -9,7 +9,6 @@ namespace Frends.SFTP.DownloadFiles.Definitions;
 public enum FileEncoding
 {
     UTF8,
-    UTF16,
     ANSI,
     ASCII,
     WINDOWS1252,